Moderate clean up:
- Remove useless Activity
- Replace custom code with Support Library
- Clean up some code
