+++ app/build.gradle
@@ -69,7 +69,6 @@ dependencies {
     compile 'com.github.KyoSherlock:expandablelayout:master-SNAPSHOT'
     compile 'com.github.pluscubed:recycler-fast-scroll:0.3.1@aar'
     compile 'com.futuremind.recyclerfastscroll:fastscroll:0.2.4'
-    compile 'de.hdodenhof:circleimageview:2.1.0'
     compile 'com.github.chrisbanes:PhotoView:1.3.0'
     compile 'com.heinrichreimersoftware:material-intro:1.6'
     compile 'su.j2e:rv-joiner:1.0.9'
