+++ app/src/main/java/be/ugent/zeus/hydra/recyclerview/viewholder/home/EventCardViewHolder.java
@@ -8,7 +8,7 @@ import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 import be.ugent.zeus.hydra.R;
-import be.ugent.zeus.hydra.activities.ActivityDetailActivity;
+import be.ugent.zeus.hydra.activities.EventDetailActivity;
 import be.ugent.zeus.hydra.models.association.Event;
 import be.ugent.zeus.hydra.models.cards.EventCard;
 import be.ugent.zeus.hydra.models.cards.HomeCard;
@@ -52,8 +52,8 @@ public class EventCardViewHolder extends HideableViewHolder {
         itemView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(itemView.getContext(), ActivityDetailActivity.class);
-                intent.putExtra(ActivityDetailActivity.PARCEL_EVENT, (Parcelable) event);
+                Intent intent = new Intent(itemView.getContext(), EventDetailActivity.class);
+                intent.putExtra(EventDetailActivity.PARCEL_EVENT, (Parcelable) event);
                 itemView.getContext().startActivity(intent);
             }
         });
