+++ app/src/main/java/be/ugent/zeus/hydra/utils/recycler/ItemSpacingDecoration.java
@@ -1,18 +1,18 @@
 package be.ugent.zeus.hydra.utils.recycler;
 
-import android.content.Context;
 import android.graphics.Rect;
+import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
 import android.view.View;
 
-import be.ugent.zeus.hydra.R;
-
 /**
- * Add spacing to elements for a staggered grid. This will add spacing on the right of every element, except for the
- * last span column.
+ * Simple item decorator that add space between elements in a {@link LinearLayoutManager}. If it is not a linear layout
+ * manager, this will do nothing.
  *
- * If the layout manager is not a {@link StaggeredGridLayoutManager}, this will do nothing.
+ * If the layout manager is in vertical mode, the space will be added to the bottom of the items.
+ * If the mode is horizontal, the spacing will be applied to the right of the items.
+ *
+ * The last item will not get spacing.
  *
  * @author Niko Strijbol
  */
@@ -20,30 +20,30 @@ public class ItemSpacingDecoration extends RecyclerView.ItemDecoration {
 
     private final int spacing;
 
-    public ItemSpacingDecoration(int spacing) {
-        this.spacing = spacing;
-    }
-
-    /**
-     * Initialize the decoration using the default spacing.
-     */
-    public ItemSpacingDecoration(Context context) {
-        this((int) context.getResources().getDimension(R.dimen.content_spacing));
+    public ItemSpacingDecoration(int mVerticalSpaceHeight) {
+        this.spacing = mVerticalSpaceHeight;
     }
 
     @Override
-    public void getItemOffsets(Rect outRect, final View view, final RecyclerView parent, RecyclerView.State state) {
+    public void getItemOffsets(Rect outRect, View view, RecyclerView parent, RecyclerView.State state) {
         super.getItemOffsets(outRect, view, parent, state);
 
-        if(parent.getLayoutManager() instanceof StaggeredGridLayoutManager) {
-            StaggeredGridLayoutManager manager = (StaggeredGridLayoutManager) parent.getLayoutManager();
+        //If this is not a LinearLayoutManager, do nothing.
+        if (!(parent.getLayoutManager() instanceof LinearLayoutManager)) {
+            return;
+        }
 
-            int spanIndex = ((StaggeredGridLayoutManager.LayoutParams) view.getLayoutParams()).getSpanIndex();
+        LinearLayoutManager manager = (LinearLayoutManager) parent.getLayoutManager();
 
-            if (spanIndex < manager.getSpanCount() - 1) {
-                outRect.right = spacing;
-            }
+        //If it is the last one, no spacing!
+        if (parent.getChildAdapterPosition(view) == state.getItemCount() - 1) {
+            return;
+        }
+
+        if (manager.getOrientation() == LinearLayoutManager.VERTICAL) {
+            outRect.bottom = spacing;
+        } else {
+            outRect.right = spacing;
         }
     }
-
 }
