+++ app/src/main/res/layout/activity_event_detail.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.ActivityDetailActivity">
+    tools:context=".activities.EventDetailActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
@@ -119,13 +119,10 @@
             android:focusableInTouchMode="true">
 
             <android.support.v7.widget.CardView
-                android:onClick="onClick"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:foreground="?android:attr/selectableItemBackground"
                 android:layout_marginBottom="@dimen/card_spacing"
                 android:layout_marginTop="@dimen/card_spacing"
-                android:clickable="true"
                 app:contentPaddingLeft="@dimen/card_text_padding_vertical"
                 app:contentPaddingRight="@dimen/card_text_padding_vertical"
                 app:contentPaddingBottom="@dimen/card_text_padding_bottom"
@@ -150,16 +147,14 @@
                         android:layout_height="wrap_content"
                         android:orientation="horizontal">
 
-                        <de.hdodenhof.circleimageview.CircleImageView
+                        <ImageView
                             android:transitionName="logo"
                             android:layout_marginTop="2dp"
                             android:id="@+id/event_organisator_image"
                             android:layout_width="40dp"
                             android:layout_height="40dp"
                             android:layout_marginRight="@dimen/vertical_padding"
-                            tools:src="@drawable/ic_image_24dp"
-                            app:civ_border_width="1dp"
-                            app:civ_border_color="@color/ugent_blue_dark"/>
+                            tools:src="@drawable/ic_image_24dp" />
 
                         <LinearLayout
                             android:layout_width="wrap_content"
