+++ app/src/main/res/layout/activity_minerva_agenda.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.ActivityDetailActivity">
+    tools:context=".activities.EventDetailActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
