+++ app/src/main/res/layout/activity_minerva_course.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.AssociationDetailActivity">
+    tools:context=".activities.minerva.CourseActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
