+++ app/src/main/res/layout/activity_resto_sandwich.xml
@@ -20,15 +20,14 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        android:animateLayoutChanges="true"
-        >
+        android:animateLayoutChanges="true">
 
         <android.support.v7.widget.RecyclerView
             android:id="@+id/resto_sandwich_recycler"
             android:scrollbars="none"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:clickable="true" />
+            app:layoutManager="@string/app_layout_manager_linear" />
 
         <com.pluscubed.recyclerfastscroll.RecyclerFastScroller
             android:id="@+id/fast_scroller"
