+++ app/src/main/res/layout/fragment_minerva.xml
@@ -54,7 +54,8 @@
         android:scrollbars="vertical"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:visibility="gone" />
+        android:visibility="gone"
+        app:layoutManager="@string/app_layout_manager_linear" />
 
     <ProgressBar
         android:paddingLeft="@dimen/vertical_padding"
