+++ app/src/main/res/layout/fragment_minerva_course_announcements.xml
@@ -1,13 +1,15 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recycler_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:listitem="@layout/item_minerva_course" />
+        tools:listitem="@layout/item_minerva_course"
+        app:layoutManager="@string/app_layout_manager_linear" />
 
     <include layout="@layout/x_progress_bar" />
