+++ app/src/main/res/values-sw600dp/recyclerview.xml
