+++ app/src/main/res/values-sw900dp/recyclerview.xml
