+++ app/src/main/res/values/recyclerview.xml
@@ -4,6 +4,8 @@
     <integer name="span_count_normal">1</integer>
     <integer name="span_count_large">1</integer>
 
+    <string name="app_layout_manager_linear">android.support.v7.widget.LinearLayoutManager</string>
+
     <string name="app_layout_manager_normal">android.support.v7.widget.LinearLayoutManager</string>
     <string name="app_layout_manager_large">android.support.v7.widget.LinearLayoutManager</string>
